Implemented BA hack
Implemented checkbox to set mesh bone node ids and mesh ids to 0 or to manually set them via the custom user properties:
CustomMeshId
CustomNodeId
Remember to add a new line after typing these, as always.
For reference, the game splits models on export to only have one material and use less than or equal to 16 bones each. The mesh id is calculated from these splits while the node id is assigned based on the dummy node created in the skeleton for each true mesh in the scene.